Changed slider updating
Altered the slider creation so the visible calculated ideology is always seen. It is updated in real time whenever new input is received. Currently the calculate button tests AI capabilities and nothing else.
