+++ calculateButton.js
@@ -6,21 +6,7 @@ import { searchGovernmentIdeologies } from "./Ideologies/GovernmentIdeologies.js
 import { GoogleGenerativeAI } from "@google/generative-ai";
 
 const calculateButton = document.getElementById("calculate");
-const economicResult = document.getElementById("economicPolicyResult");
-const foreignResult = document.getElementById("foreignPolicyResult");
-const governmentResult = document.getElementById("governmentPolicyResult");
 
-calculateButton.addEventListener("click", () => {
-  let currResult = searchEconomicIdeologies(getScore()[0]);
-  economicResult.textContent = currResult.name + ": " + currResult.description;
-
-  currResult = searchForeignIdeologies(getScore()[1]);
-  foreignResult.textContent = currResult.name + ": " + currResult.description;
-
-  currResult = searchGovernmentIdeologies(getScore()[2]);
-  governmentResult.textContent =
-    currResult.name + ": " + currResult.description;
-});
 // WARNING: Storing API keys directly in client-side code is INSECURE for production.
 const API_KEY = "AIzaSyB61E3oz5jQwABufrXxc4klURp7mOA2AVQ";
 
@@ -34,4 +20,6 @@ async function generateContent() {
   console.log(response.text());
 }
 
-generateContent();
+calculateButton.addEventListener("click", () => {
+  generateContent();
+});
